[feat](index table): use antd's table to add sorter
